% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{add_names}
\alias{add_names}
\title{Make an iterator returns named values.}
\usage{
add_names(itr, value)
}
\arguments{
\item{itr}{an iterator}

\item{value}{a character vector of desired names}
}
\value{
an iterator
}
\description{
This function is significantly slower since names are set to returned objects
 everytime the iterator is called.
}
\examples{
itr <- iterator_product(1:3, 4:5)
itr_named <- add_names(itr, c("a", "b"))
collect(itr, 5)
}
