% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom.R
\name{CleanupGenerator}
\alias{CleanupGenerator}
\title{Stop a non-blocking generator worker process and cleanup.}
\usage{
CleanupGenerator(generator)
}
\arguments{
\item{generator}{a generator function.}
}
\value{
last result from generator(), invisibly.
}
\description{
Because a generator loops indefinitely, when invoking non-blocking generators,
child processes are always forked in background loading next batches. Even though
the generators are not need anymore. Calling this function results in stopping
the child process and any consequence IO.
}
\details{
This function force stops the multi-process generator and remove the function
from its parent frame to make gc works.
}
