% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_aux.R
\name{collect}
\alias{collect}
\title{Collect multiple iterate values from an iterator.}
\usage{
collect(itr, n = NULL, simplify = TRUE)
}
\arguments{
\item{itr}{an iterator or a counter function.}

\item{n}{how many iterates to collect.}

\item{simplify}{passed to sapply()}
}
\value{
a vector/matrix/list depending on iterator.
}
\description{
Collect multiple iterate values from an iterator.
}
\examples{
itr <- iterator_zip(1:3, 6:8)
collect(itr, 5)
}
