% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{CountClass}
\alias{CountClass}
\title{Count number of samples of each class.}
\usage{
CountClass(onehot, ret.density = FALSE)
}
\arguments{
\item{onehot}{a onehot table.}

\item{ret.density}{wheter to return density, otherwise number of samples.}
}
\value{
a numeric vector.
}
\description{
Count number of samples of each class.
}
