% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{counter}
\alias{counter}
\title{Create a counter that counts forever (before overflowing).}
\usage{
counter(start = 1L, step = 1L)
}
\arguments{
\item{start}{counter start}

\item{step}{step size}
}
\value{
a counter function.
}
\description{
Create a counter that counts forever (before overflowing).
}
\examples{
ctr <- counter(1, .5)
collect(ctr, 10)
}
