% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom.R
\name{DataGeneratorNonblock}
\alias{DataGeneratorNonblock}
\title{Non-blocking data generator.}
\usage{
DataGeneratorNonblock(loader, sample_ids, sample_cls = NULL, batch_size)
}
\arguments{
\item{loader}{a loader function that accepts a subset of sample_ids and returns
loaded data.}

\item{sample_ids}{a vector of ids passed to loader.}

\item{sample_cls}{OPTIONAL, a vector of classes returned as y.}

\item{batch_size}{an integer of batch size.}
}
\value{
a generator function.
}
\description{
Non-blocking data generator.
}
