% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{DecodeOnehot}
\alias{DecodeOnehot}
\title{Decode a onehot table to a list of classes}
\usage{
DecodeOnehot(onehot, value_offset)
}
\arguments{
\item{onehot}{a onehot table.}

\item{value_offset}{the offset value used to encode onehot table.}
}
\value{
a list of integer classes
}
\description{
Decode a onehot table to a list of classes
}
