% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{EncodeClassList}
\alias{EncodeClassList}
\title{Encode a list of classes to a onehot table.}
\usage{
EncodeClassList(cls_list, nclass = NULL, value_offset = NULL)
}
\arguments{
\item{cls_list}{a list of integer classes.}

\item{nclass}{OPTIONAL, number of classes.}

\item{value_offset}{OPTIONAL, a value to offset smaller than 1 values in cls_list.}
}
\value{
a onehot table.
}
\description{
Encode a list of classes to a onehot table.
}
