% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_png.R
\name{EstimateMem}
\alias{EstimateMem}
\title{Estimate memory usage of images.}
\usage{
EstimateMem(n, h, w, nch, unit = "k")
}
\arguments{
\item{n}{number of images to load.}

\item{h}{height of image.}

\item{w}{width of image.}

\item{nch}{number of channels.}

\item{unit}{what unit to use, "k" for KB, "m" for MB and "g" for GB,
otherwise bytes.}
}
\value{
a numeric.
}
\description{
Estimate memory usage of images.
}
