% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{fake_vectorise}
\alias{fake_vectorise}
\title{Fake vectorisation of function}
\usage{
fake_vectorise(f, ...)
}
\arguments{
\item{f}{a non-vectorised function}

\item{...}{additional arguments passed to wrapped function}
}
\value{
a function
}
\description{
Fake vectorisation of function
}
\examples{
f <- function(x) length(x)
f2 <- fake_vectorise(f)
f(c(1,2,3,4)) #4
f2(c(1,2,3,4)) #[1,1,1,1]
}
