% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{IdxToClass}
\alias{IdxToClass}
\title{Get hot classes from indices}
\usage{
IdxToClass(idx, onehot, f_reduce = NULL, ret.sorted = FALSE, ...)
}
\arguments{
\item{idx}{a vector of indices.}

\item{onehot}{a onehot table.}

\item{f_reduce}{function to reduce, if NULL, union is performed.}

\item{ret.sorted}{return sorted classes.}

\item{...}{arguments passed to f_reduce.}
}
\value{
a vector of classes.
}
\description{
Get hot classes from indices
}
