% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_aux.R
\name{is.iterator}
\alias{is.iterator}
\title{Test if an object is an iterator.}
\usage{
is.iterator(x)
}
\arguments{
\item{x}{an object.}
}
\value{
TRUE or FALSE
}
\description{
Test if an object is an iterator.
}
\examples{
is.iterator(counter(1, 3)) #not an iterator
is.iterator(iterator_atomic(1:3))
}
