% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator_atomic}
\alias{iterator_atomic}
\title{Create iterator over an atomic vector.}
\usage{
iterator_atomic(x)
}
\arguments{
\item{x}{an iterable atomic vector.}
}
\value{
an iterator.
}
\description{
Create iterator over an atomic vector.
}
\examples{
itr <- iterator_atomic(3:5)
#Created iterator loops indefinitely.
print(collect(itr, 10))
itr <- iterator_atomic(c("a", "b", "c", "d"))
itr()
}
