% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{iterator_atomic_cond}
\alias{iterator_atomic_cond}
\title{Conditioned atomic iterator}
\usage{
iterator_atomic_cond(x, cond)
}
\arguments{
\item{x}{an atomic vector}

\item{cond}{a conditional function, must be vectorised.}
}
\value{
an iterator
}
\description{
The conditions are evaluated only once thus cond() should NOT have side effects.
}
\examples{
x <- runif(n = 10)
f <- function(a) a < 0.8
itr1 <- iterator_atomic(x)
itr2 <- iterator_atomic_cond(x, f)
collect(itr1)
collect(itr2)
}
