% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator_batch}
\alias{iterator_batch}
\alias{iterator_partition}
\title{Iterate over batch_size elements a time.}
\usage{
iterator_batch(x, batch_size = 1L)

iterator_partition(x, part_size = 1L)
}
\arguments{
\item{x}{an iterable atomic vector.}

\item{batch_size}{batch size.}

\item{part_size}{partition size.}
}
\value{
an iterator.
}
\description{
Iterate over batch_size elements a time.
}
\examples{
itr <- iterator_batch(500:5000, 32)
itr()
}
