% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{iterator_batch_cond}
\alias{iterator_batch_cond}
\alias{iterator_partition_cond}
\title{Conditioned batch iterator}
\usage{
iterator_batch_cond(x, batch_size = 1L, cond)

iterator_partition_cond(x, part_size = 1L, cond)
}
\arguments{
\item{x}{an atomic vector}

\item{batch_size}{batch size}

\item{cond}{a conditional function, must be vectorised.}

\item{part_size}{partition size.}
}
\value{
an iterator
}
\description{
Conditioned batch iterator
}
\examples{
x <- runif(n = 50)
f <- function(a) a < 0.8
itr1 <- iterator_batch(x, 3)
itr2 <- iterator_batch_cond(x, 3, f)
collect(itr1)
collect(itr2)
}
