% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator_product}
\alias{iterator_product}
\title{Iterate over a list of iterators or atomic vectors. Cartesian product style.}
\usage{
iterator_product(...)
}
\arguments{
\item{...}{iterators or atomic vectors.}
}
\value{
an iterator.
}
\description{
Iterate over a list of iterators or atomic vectors. Cartesian product style.
}
\examples{
itr1 <- iterator_atomic(1:3)
itr <- iterator_product(itr1, 1:2)
collect(itr, 10)

itr1 <- iterator_atomic(1:3)
itr2 <- iterator_atomic(1:2)
itr3 <- iterator_product(itr1, itr2)
itr <- iterator_product(itr3, 5:8)
collect(itr, 20)
}
