% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator_zip}
\alias{iterator_zip}
\title{Iterate over a list of iterators or atomic vectors. Zip style.}
\usage{
iterator_zip(...)
}
\arguments{
\item{...}{iterators or atomic vectors.}
}
\value{
an iterator.
}
\description{
Iterate over a list of iterators or atomic vectors. Zip style.
}
\examples{
itr1 <- iterator_atomic(1:6)
itr <- iterator_zip(itr1, 7:12)
collect(itr, 10)

#A warnning will be thrown if size of iterators are different.
itr1 <- iterator_atomic(1:6)
itr2 <- iterator_atomic(1:4)
itr <- iterator_zip(itr1, itr2)
collect(itr, 15)
}
