% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R, R/images_png.R
\name{LoadImage}
\alias{LoadImage}
\alias{LoadPNG}
\title{Load images and concatenate channels.}
\usage{
LoadImage(files, loader, h = NULL, w = NULL, nch = NULL,
  channel_last = TRUE, ret.flatten = FALSE)

LoadPNG(files, h = NULL, w = NULL, nch = NULL, channel_last = TRUE,
  ret.flatten = FALSE)
}
\arguments{
\item{files}{files to load.}

\item{loader}{a function to load the images.}

\item{h}{OPTIONAL, height of the image. Default: Evaluated during run.}

\item{w}{OPTIONAL, width of the image. Default: Evaluated during run.}

\item{nch}{OPTIONAL, total channel number of the concatenated image. Default: Evaluated during run.}

\item{channel_last}{whether the images are channel_last format. Default: TRUE}

\item{ret.flatten}{whether to return a flatten vector. Default: FALSE}
}
\value{
a 3-d array.
}
\description{
Load images using the given loader() function, and concatenate all images
loaded as a new multi-channel image.
}
\details{
It is encouraged to provide h and w unless ret.flatten is TRUE, since an extra
loading is run to evaluate them thus reducing performance.
}
