% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R, R/images_png.R
\name{LoadImageBatch}
\alias{LoadImageBatch}
\alias{LoadPNGBatch}
\title{Load a batch of images.}
\usage{
LoadImageBatch(files, loader, h, w, nch, channel_last = TRUE,
  batch_last = FALSE)

LoadPNGBatch(files, h, w, nch, channel_last = TRUE, batch_last = FALSE)
}
\arguments{
\item{files}{a list of files or a list of files by channel.}

\item{loader}{a function to load the images.}

\item{h}{height of images.}

\item{w}{width of images.}

\item{nch}{number of channels of each image.}

\item{channel_last}{whether the images are channel_last format. Default: TRUE}

\item{batch_last}{whether the batch is indexed in last dimension. Default: FALSE}
}
\value{
a 4-d array.
}
\description{
Load a batch of images.
}
\details{
The argument files can be either a list of files to load, or a list
of which each element contains a list of files to load and concatenate by
channel.
The latter can be handy when data is provided in single-channel image files.
}
