% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_png.R
\name{LoadPNGChan}
\alias{LoadPNGChan}
\title{Load and concatenate png channels.}
\usage{
LoadPNGChan(channels, h = NULL, w = NULL, ret.flatten = FALSE)
}
\arguments{
\item{channels}{file names for each channel.}

\item{h}{OPTIONAL, height of the image.}

\item{w}{OPTIONAL, width of the image.}

\item{ret.flatten}{return a flatten vector.}
}
\value{
a channel last 3-d array.
}
\description{
Channel last is assumed. Use aperm to convert to channel first if needed.
}
\details{
If h or w is not given, an extra loading/evaluation on first channel is done
on first channel unless ret.flatten, so some performance loss I guess.
}
