% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_png.R
\name{LoadPNGList}
\alias{LoadPNGList}
\title{Load a list of png.}
\usage{
LoadPNGList(files, h, w, nch, file_idx_first = TRUE)
}
\arguments{
\item{files}{a list of filenames or a list of channel filenames.}

\item{h}{height of image.}

\item{w}{width of image.}

\item{nch}{total channels of image.}

\item{file_idx_first}{whether to put file index in fisrt dimension, otherwise it's put in last dim.}
}
\value{
a 4-d array.
}
\description{
This function is essentially an lapply of LoadPNGChan on files.
}
