% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{names<-.iterator}
\alias{names<-.iterator}
\title{Make an iterator return named values.}
\usage{
\method{names}{iterator}(x) <- value
}
\arguments{
\item{x}{an iterator}

\item{value}{a character vector of desired names}
}
\value{
an iterator
}
\description{
Make an iterator return named values.
}
\details{
Since this is just a wrapper of iterator, renaming an iterator means
additional wrapped function call which adds to overhead thus worse performance.
}
\examples{
itr <- iterator_product(1:3, 4:5)
names(itr) <- c("a", "b")
collect(itr, 5)
}
