% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator2.R
\name{names.iterator}
\alias{names.iterator}
\title{Element names of an iterator}
\usage{
\method{names}{iterator}(x)
}
\arguments{
\item{x}{an iterator.}
}
\value{
a character or NULL
}
\description{
Element names of an iterator
}
\examples{
itr <- iterator_product(1:3, 4:5)
print(names(itr))
names(itr) <- c("a", "b")
print(names(itr))
}
