% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuity.R
\name{pad_idx_left}
\alias{pad_idx_left}
\alias{pad_idx_right}
\alias{pad_idx}
\title{Pad indices to left or right}
\usage{
pad_idx_left(idx, start = 1L)

pad_idx_right(idx, end = NA)

pad_idx(idx)
}
\arguments{
\item{idx}{a vector of indices}

\item{start}{left most index}

\item{end}{right most index}
}
\value{
a vector of padded indices
}
\description{
Pad indices to left or right
}
\examples{
n = 1000
xi <- sort(sample(seq(n), 100))
xleft <- pad_idx_left(xi)
xright <- pad_idx_right(xi, n)
}
