% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{RelatedIdx}
\alias{RelatedIdx}
\title{Find all related indices.}
\usage{
RelatedIdx(idx, onehot, ...)
}
\arguments{
\item{idx}{a vector of indices.}

\item{onehot}{a onehot table.}

\item{...}{arguments passed to IdxToClass and ClassToIdx.}
}
\value{
a vector of indices.
}
\description{
By related indices I mean all indices that share a same class.
}
