% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discontinuity.R
\name{sample_idx}
\alias{sample_idx}
\title{Indices of samples}
\usage{
sample_idx(measured, sampled, cmpf = `>=`)
}
\arguments{
\item{measured}{a vector of measured values, must be sorted by cmpf}

\item{sampled}{a vector of sampled values, must be sorted by cmpf}

\item{cmpf}{a test function to compare sampled values to measured values}
}
\value{
a vector of indices of measured values
}
\description{
this niche function compare sampled values to measured values
cmpf(), and returns a vector of first passed indices for every sampled value.
The assumption of measured and sampled being sorted by cmpf guarantees cmpf
is called only N times provided length(measured) == N.
}
\examples{
measured <- sort(sample(seq(1000000), 500000))
sampled <- sort(sample(measured, 20000) + runif(20000, -100, 0))
indices <- sample_idx(measured, sampled, cmpf = `>=`)
stopifnot(all(measured[indices] >= sampled))
}
