% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{ScarceClass}
\alias{ScarceClass}
\title{Find classes that is scarce.}
\usage{
ScarceClass(threshold, onehot)
}
\arguments{
\item{threshold}{a threshold, any number of idx < threshold will return.}

\item{onehot}{a onehot table.}
}
\value{
an integer vector.
}
\description{
Find classes that is scarce.
}
