% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_aux.R
\name{size}
\alias{size}
\title{Size of an iterator.}
\usage{
size(itr)
}
\arguments{
\item{itr}{an iterator.}
}
\value{
an integer size.
}
\description{
Size of an iterator.
}
\examples{
itr <- iterator_product(1:3, 5:6, 2)
size(itr)
}
