% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{SwapMissingClass}
\alias{SwapMissingClass}
\title{Swap portion of the missing classes from idx_ref to idx.}
\usage{
SwapMissingClass(idx, idx_ref, onehot, swap_ratio = 0.1,
  swap_random = FALSE)
}
\arguments{
\item{idx}{a vector of indices.}

\item{idx_ref}{a vector of indices.}

\item{onehot}{a onehot table.}

\item{swap_ratio}{ratio of swapped elements from idx_ref.}

\item{swap_random}{whether swapped elements are randomly selected by sample(),
otherwise elements are selected in index order.}
}
\value{
a named list with new idx and idx_ref.
}
\description{
Could be useful to re-construct idx and idx_ref with no missing classes. Set
swap_ratio to 0 results in a relatively (but not globally) minimal swapping.
}
