% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{ClassToIdx}
\alias{ClassToIdx}
\title{Get hot indices from classes.}
\usage{
ClassToIdx(cls, onehot, f_reduce = NULL, ret.sorted = FALSE,
  fast = TRUE, ...)
}
\arguments{
\item{cls}{a vector of classes.}

\item{onehot}{a onehot table.}

\item{f_reduce}{function to reduce, if NULL, union is performed.}

\item{ret.sorted}{return sorted indices.}

\item{fast}{whether to use faster algorithm when f_reduce is NULL, result is
always sorted in this case.}

\item{...}{arguments passed to f_reduce.}
}
\value{
a vector of indices.
}
\description{
Get hot indices from classes.
}
