% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom.R
\name{CleanupGenerator}
\alias{CleanupGenerator}
\title{Stop a non-blocking generator worker process and cleanup.}
\usage{
CleanupGenerator(generator)
}
\arguments{
\item{generator}{a generator function.}
}
\value{
last result from generator(), invisibly.
}
\description{
This function might block since it waits for last worker to return. It has the
same effect of calling generator(STOP = TRUE) but returns value invisibly with
an extra call to gc().
}
