% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{EncodeClassList}
\alias{EncodeClassList}
\title{Encode a list of classes to a onehot table.}
\usage{
EncodeClassList(cls_list, nclass = length(unique(unlist(cls_list))),
  value_offset = 1L - min(unlist(cls_list)))
}
\arguments{
\item{cls_list}{a list of integer classes.}

\item{nclass}{number of classes.}

\item{value_offset}{a value to offset smaller than 1 values in cls_list.}
}
\value{
a onehot table.
}
\description{
Encode a list of classes to a onehot table.
}
