% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{MissingClass}
\alias{MissingClass}
\title{Find missing classes in a set of indices compared to reference indices.}
\usage{
MissingClass(idx, idx_ref, onehot)
}
\arguments{
\item{idx}{a vector of indices.}

\item{idx_ref}{a vector of indices.}

\item{onehot}{a onehot table.}
}
\value{
a vector of classes.
}
\description{
Find missing classes in a set of indices compared to reference indices.
}
