% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{RelatedClass}
\alias{RelatedClass}
\title{Find all related classes.}
\usage{
RelatedClass(cls, onehot, ...)
}
\arguments{
\item{cls}{a vector of classes.}

\item{onehot}{a onehot table.}

\item{...}{arguments passed to IdxToClass and ClassToIdx.}
}
\value{
a vector of classes.
}
\description{
By related classes I mean all classes that share a same index.
}
