% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot.R
\name{RelatedIdx}
\alias{RelatedIdx}
\title{Find all related sample indices.}
\usage{
RelatedIdx(idx, onehot, ...)
}
\arguments{
\item{idx}{a vector of sample indices.}

\item{onehot}{a onehot table.}

\item{...}{arguments passed to IdxToClass and ClassToIdx.}
}
\value{
a vector of indices.
}
\description{
Related samples are those contain same classes.
}
